Add tagging to resources created
We primarily have a consistent pattern for Name, Owner, and Contact across different resource types, so started there. Added these as required parameters to make it easier for both those who use the aws cli or the web console to populate.These will likely need to be changed a bit once we have a better understanding of how other institutions may use these.