ESU-675: Bugfix pipeline name
Changed references to the stack name to use the given pipeline name. This is not as friendly a name, but is also a little more accurate and less dependent on stack naming conventions.

An extra, unrelated change is to the default desired count of services to run for the image service. We are currently in a pattern of only ever using one, so the default was going against our current pattern.