Add DNS to manifest pipeline
This change adds the ability to optionally create a zone in Route53, and a record for the manifest-pipeline and unified-static-host stacks, in order to better automate the creation of a valid hostname and certificate. Summary of changes:
- Changed `manifest-pipeline.yml` and `unified-static-host.yml` to allow for the optional create of a RecordSet in Route53
- Separated the creation of the ACM and zone into a `domain.yml`. This can optionally skip the creation of the zone to only create the wildcard cert for the domain/subdomain. Note: this still assumes we’re continuing with a single wildcard cert for each subdomain for now. ACM was previously created in app-infrastructure, but this will allow for creating separate subdomains for different grouping of services, ex: `*.domain.com` and `*.iiif.domain.com`.
- Changed all of the lower level templates to import the domain name and cert arn from the new domain stack
- Changed the manifest-pipeline-pipeline to propagate down to the lower level stacks for several params: DomainStackName, SubDomain for both test/prod, and the optional flag for CreateDNSRecord. This means both adding an input param to the template, then adding it to the config parameters in the build stage.
- Added permissions to the CloudFormation role to perform the necessary operations on DNS
- Added an output in the `manifest-pipeline.yml` to make it easier to get the CloudFront dns name, for when someone is manually creating the CNAME in external DNS
- Updated README to add the new domain stack and changed others to add new required parameters. Also moved the manifest pipeline example out of the notifications section

Unrelated to DNS, I also added a max length to the manifest stack names in `manifest-pipeline-pipeline.yml`. There is a naming issue caused by the fact that we need to allow CloudFormation to name things and, because we don't know what those names will be precisely, we must use wildcard patterns on the service role for the pipeline. When a name is not given, CloudFormation will name things as follows `<(truncated)stack name>-<(truncated)resource identifier>-<random id>`. Depending on the length of the resource identifier + stack name, either your stack name, the identifier, or both can get truncated. This will cause problems with the wildcards in the IAM policies because they are using the full (non-truncated) stack name. I did some experimentation on this and found that if both are longer than 24 characters, it will truncate them both, so I'm limiting stack name to 24. In the past, I attempted to handle this by explicitly naming things based on stack names (see #38), but this is proving to cause additional issues with lambdas, especially the new edge lambda. When updating the stack it will require replacement (see FunctionName in https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-lambda-function.html), which is problematic. So, I've reverted all of the explicit naming changes that were in #38, except for the StepFunction, and added the stack name limit. This may or may not solve all of the issues here, so we'll have to keep monitoring this.