Bug fixes related to rebuild
During yesterday's deploy, encountered a few issues with building everything back up. The image service pipeline was unable to read from SSM for the prod host name when posting to github. Additionally, when deploying without creating the dns zone, the domain stack will not have a Zone export. This currently breaks the manifest code pipeline. Had to remove this dependency to get it working. We need to revisit this, but the best solution I could determine to work with both paths is to just allow all. This should not be a security issue in our production account for now, since the code pipeline is not creating DNS record sets, but we'll want to fix this before we redeploy to our development account.