Group Approval with Test deployment stages
This changes all pipelines to ensure that any QA and Approval steps are grouped within the same stage as the Deploy to the test environment. This will prevent an integrity issue that can occur if multiple invocations of the pipeline happen before the Approver can visit the test site.
- Changed IIIF service deployment pipeline to group Deploy/QA/Approval into one stage.
- Changed static-host-pipeline to group the deploy/approval actions under a test stage

A few additional cleanups/fixes:
- Found a new reference to mellon in the README and changed to marble
- Changed default production stack names to include a -prod suffix to better match how we are now deploying these stacks
- Renamed several stages and actions to be more consistent across the different pipelines
- Fixed an issue with redeploying the lamba in manifest-passthrough-pipeline