Make requiring MFA optional
This changes the MFA condition on the shared user roles to be optional. I also changed the condition to be a straight Bool check instead of an IfExists check. The aws:MultiFactorAuthPresent key seems to always be there in the contexts that I checked, such as cli, console, and federated user console. So, it's unclear when this key is not present, but given the way this template parameter reads we wouldn't want it to ignore the condition in certain contexts if the expectation would be to always deny when MFA was not used. Updated the shared-infrastructure readme to include info about this parameter and how to use with federated logins.