Change node version in static-host-pipeline
This bumps the build to use the latest nodejs. It also changes the CodeBuild to use the new polyglot way of defining the environment (https://aws.amazon.com/about-aws/whats-new/2019/07/aws-codebuild-adds-support-for-polyglot-builds/). I've tested this on both the website and image-viewer builds. Both seem to work, although image-viewer has a lot of warnings that did not exist in prior executions. We may need to either parameterize this in the future or update the image-viewer packages.