Fix Image Service access logs
This is a hotfix to get the deployment to work. This adds a bucket policy to the shared log bucket to allow the ALB to write access logs. For now this will only work when the image service is deployed to us-east-1. This additionally extends the timeout between the ALB and Cantaloupe based on testing performed when rendering large images. This will likely need to be reassessed once we determine the sizes we will support.