Create deployment for nulib/iiif-serverless
This adds a deployment pipeline that will deploy the app using the provided template.yml in the app repo. This also adds a new CloudFormation that will create a CustomDomain for the API so that our hostnames and certs will match. Creating this in cdk would have made an odd mix between the two. In the future, we'll want to refactor to use a cdk based stack for the application as well, so I'm assuming it will be easier to rewrite this part at that time.

I'm putting this cdk code at the root of the deploy/cdk directory. Personally, I think this should be a multistack cdk application so that on deploy you can just choose which components (stacks) you want to deploy. In addition, doing it as a multistack application will let cdk take care of a lot of the cross-stack property sharing for us (although it will be via export/import only atm). Unfortunately, this will look a little confusing for a bit since we have some applications within subdirectories for the other components (elasticsearch and web-kiosk).