Override create DNS record
Need a way to ensure this does not create DNS in our production account. Had to add some top level context entries so that we can more easily override these with `-c` on the cli when deploying within CodeBuild. Also renamed the stacks to match what we previously did for the IIIF service