Add user-content deployment pipeline
This adds a deployment pipeline for the user-content component. Changes:
- Introduced a new namespace context as a global value to prefix names
with. The expectation will be that all cdk based stacks will be prefixed
with this namespace so that we can differentiate different components
and environments. For example, a developer can independently deploy a
user content stack by specifying a namespace of `-c namespace=my-marble`
and then specifying the stack to deploy as `my-marble-user-content`.
Likewise the pipeline will adhere to this namespace, ex: when deploying
`my-marble-user-content-deployment`, the pipeline will deploy the test
stack as `my-marble-test-user-content` and the prod stack as
`my-marble-prod-user-content`. This is somewhat counter to what we
currently do for environment, but the context overrides are already
getting out of hand, so I'd prefer we minimize this by just doing a
single namespace param.
- Created, what I hope will be, a reusable class for handling cdk
eployments in a pipeline. The class will handle common things such as
handling the cdk install in the buildspec, and give it permissions
needed to access the cdk bucket and to change the designated stacks
- Changed the cdk project to use yarn
- Added slack approvals to the pipeline as an optional add, by
specifying `-c slackNotifyStackName`. Requires using our slack approval
bot https://github.com/ndlib/codepipeline-approvals/blob/master/slack_approval.md

User content component changes:
- Moved UserContentStack to its own file and changed index so that we
can export multiple classes for this component
- Changed to let cloudformation name the tables to avoid update failures
- Added readme for deploying user-content stack and pipeline stack

Namespace policy changes:
- The user-content component adds a `AWS::ApiGateway::Account` resource,
which requires api:PATCH to `arn:aws:apigateway:${AWS::Region}::/account`
So, added this to the api permissions
- This component also uses dynamodb, so added a helper function for
giving permissions to perform actions on dynamodb things based on the
stack name

Context changes:
- Using objects within context is just going to be too hard to use when
overriding within a pipeline, so flattened the userContent keys for now.
Will want to update the iiif context to do this as well.
- Apparently cdk wants context to be in the separate file now, so moved