MARBLE-678 Change Unique Identifier for Users
* Previously the unique identifier for users was the user `netid`. This 
would become problematic when we implement alternative login methods 
such as Facebook or Google later on. To remedy this the unique 
identifier will be changed to a hash value of the `sub` and `iss` fields 
on the user token. Most of the needed change will take place in the 
Marble User Content lambda code, but we need an additional `GET` 
endpoint for the API and an additional secondary global index on the 
user table to continue being able to look up users by `userName`.