Add options to use existing VPC and/or DNS HostedZone
- Added option to import an existing Vpc by specifying a useVpcId context
variable
- Added option to import an existing HostedZone by specifying a
useExistingHostedZone context variable. When true, the hosted zone will
be looked up by the domainName context variable.
- Added helper for mocking out Vpc.fromLookup and HostedZone.fromLookup
calls and changed jest config to restore mocks after each test to ensure
integrity.
- DRY'd up tests a bit more

Context and environment agnostic related changes and rationale

You will see that I'm now git ignoring the cdk.context.json, which deserves
a bit of an explanation. We are trying to make env agnostic code, but
that can be at odds with the way cdk wants to do things. For example,
I wanted to make the new options as easy to use as possible. So, instead
of requiring a user to input VPC, subnets, gateway ids, etc as we did in
CloudFormation, I wanted cdk to do this for us via the `fromLookup`
factory methods. The first issue with this is that [env](https://docs.aws.amazon.com/cdk/latest/guide/environments.html) must be set for the
stacks for these to work, making it less agnostic. The second issue, is
that when you use an existing vpc, then share that to dependent stacks,
no export/import binding is created. This means that dependent stacks
must be deployed and given the same useVpcId option that was used when
the foundation stack is deployed. Because of these issues, I see no other
option than to begin using context the way it was meant to be used so that
we can persist values (such as vpc lookups) so that all developers will
deploy to the account in the same way.

In order to do this while still making it usable in a somewhat agnostic
manner, I've opted to ignore the cdk.context.json and instead provide an
example config, a config that our developers will use in dev account, and
(eventually) a config that we will use in our production account. The gist
is a new user will have to copy one of these configs to cdk.context.json
before running any cdk commands. The copy/paste is necessary because cdk
does not allow [specifying a context file at deploy time](https://github.com/aws/aws-cdk/issues/3760). If a new key
is added to the context, we will need to add it to all three files.