Moving environment into context
Created a subobject in context so that we can specify values that are
specific to an environment such as Vpc ids. Created a dev and prod
environment for our two accounts. There may be some other things that are
specific to environment such as the image processing pipeline that we'll
need to also refactor.

I tested this with the user-content deployment pipeline, and aside from
having to fix some new permission issues that cropped up unrelated to
these changes, it worked fine. The only extra bit that I had to do was
deploy a marble-test-foundation and a marble-prod-foundation since the
pipelines are exclusively deploying their components yet now expect one
of these for both test and prod.