Change namespaced context handling
We have several key value pairs in our context that are directly mapped
to stack props. These kvp are namespaced as "namespace:key". I found a
way to get all context from an env variable so that we can just glob all
of these and pass them into the stacks. This should make it cleaner and
easier to add new context/props to stacks.