Add IIIF Serverless API
This adds the ability to deploy the serverless-iiif stack with cdk. This
implementation still makes use of the existing template.yml in the source
repository. This meant having to work around a few issues with doing this
in cdk. Alternatives would be to just create the API via cdk code and
remove our dependence on the template.yml or deploying straight from the
SAR application that nulib publishes, both of which would require
discussion.

This does not completely fix the pipeline yet.

This depends on https://github.com/ndlib/serverless-iiif/pull/5