Created Manifest Deployment in CDK
I added several parameters to cdk-context.json which are required by this stack - to minimize required supplied parameters when deploying.

One thing I didn't include in this migration, which was included in the previous cloud formation, was a post deploy step to update the Application github repository to indicate this had been deployed.