Separate services from pipelines
Developing and deploying cd pipelines is a different type of activity
from developing and deploying the service stacks. In addition to this,
we are now creating separate foundation stacks for test and prod, but not
specifically instantiating these stacks in the cdk app, which is an
anti-pattern. Putting all of these things in base.ts was becoming harder
to parse and understand. This is an attempt at separating these types of
activities. The idea is that you will now have to pass an additional
context variable called `stackType`. The available stacks to deploy will
depend on what is given for this context variable, either `service` or
`pipeline`. `service` will be the default to retain backward compatibility
with both devs and pipelines.

Here's the output of `cdk list` as an example:
```
$ cdk list -c env=dev
marble-elastic
marble-foundation
marble-image-processing
marble-image-service
marble-manifest
marble-redbox
marble-user-content
marble-website

$ cdk list -c env=dev -c stackType=pipeline
marble-image-processing-deployment
marble-prod-elastic
marble-prod-foundation
marble-test-elastic
marble-test-foundation
marble-user-content-deployment
marble-image-service-deployment
marble-prod-image-processing
marble-prod-image-service
marble-prod-manifest
marble-prod-redbox
marble-prod-user-content
marble-prod-website
marble-redbox-deployment
marble-test-image-processing
marble-test-image-service
marble-test-manifest
marble-test-redbox
marble-test-user-content
marble-test-website
marble-website-deployment
```

With this change, if your pipeline needs to reference any resources from
the foundation stack, you'll need to pass both the test and prod instances.
I've converted the existing ones to do this, and with this was also able
to add a stage in the IIIF pipeline that will copy the test images into
the target test/prod buckets before performing tests, making the entire
pipeline much easier to prop up from scratch.