Cloudformation cleanup
The core components have all been rewritten in cdk. There may still be
some bug fixes that we'll find we need as we try to redeploy everything
anew, but at this point we should be able to safely remove these templates.
In addition, the sentry layer is no longer used, so removing.