DRYing up context a bit
- Changed all pipelines to read the infra repo settings from a single set
of shared kvp from context.
- Split up the context kvp for website and redbox and added values for
the hostnames so that things are more statically defined
- Removed QA repo source from static host pipeline. These are all just
using the app repo for qa tests

With these changes we should be able to simply run something like
`cdk deploy marble-*-deployment -c env=dev -c stackType=pipeline`
to deploy all of the CD pipelines. We can also more easily override the
branch we source from the infra repo with a single override for all of
the pipelines with `-c infraSourceBranch=my-branch`