Fix issues in image processing CD pipeline
- Changed image processing pipeline to get manifest pipeline bucket names
from other stacks instead of requiring them via context
- Changed image processing to use the env config to get the rbsc bucket
since that changes per env
- Fixed an issue with the target stack names. We had changed them from
`marble-image` to `marble-image-processing` but had missed updating the
pipeline to use these names when deploying.
- Fixed an issue with the code terminating too soon during synth,
preventing cdk from seeing dependencies correctly. We need to find an
alternative solution to the `if !fs.existsSync(props.lambdaCodePath); return`
checks that still allow the code to continue, but still throws an error
when that stack is deployed. In this specific case, cdk was not seeing
the dependency between image processing and manifest pipeline because
the manifest pipeline was exiting too soon due to the missing files. For
now, I just moved these fs checks down a bit and forced the export/import
to happen in the manifest pipeline. This is a pretty fragile solution,
and it's not very intuitive what's happening when this occurs, so I'm
going to look for a more general solution to this, but this fixes the
immediate issue for now. See https://github.com/aws/aws-cdk/issues/6743
for the original issue we were trying to prevent with these checks.