Fix search url
This further automates deployment by sharing the ES endpoint and index
name with the rest of the app. The endpoint will now be added to ssm
by the elasticsearch stack, which will be injected into the env of
the website CodeBuild as SEARCH_URL. This alleviates the need to manually
create the `/all/static-host/marble-test-website/search_index` and
`/all/static-host/marble-test-website/search_url` parameters.