Fix manifest pipeline tests
This breaks out the unit tests for the app into a separate stage from
the test deploy. Unit tests can always be safely re-run, even if the
pipeline is stalled in a deploy stage. It also has a different set of
package dependencies from a production deploy, so it should be managed
in a separate environment.