Add better asset handling
This replaces the clunky "if asset missing, return" escapes. Returning
is bad, since it prevents synth'ing the remainder of the stack. This
adds a few helpers to use in place of Container and Lambda asset calls
that will return a dummy object if the asset is not found so that the
application can continue synth, but will add an error to the scope to
prevent allowing a deploy.

This is likely something I will add to ndlib-cdk, but adding here for
now to get past an immediate issue in CI.