Marble 1662 modify app sync to accommodate new image and media structures (#295)
Marble-1662 Modify AppSync to accommodate image and media separately
Several changes were made to maintain-metadata:
* Add types to schema: Media, MediaGroups, Image, and ImageGroups
* Add Queries: getMedia, getMediaGroup, getImage, getImageGroup, listImageGroups, listImageGroupsByStorageSystem, listImageGroupsForS3, listMediaGroups, listMediaGroupsByStorageSystem, listMediaGroupsForS3.
* Add mutation saveImageGroupForWebsite
* Modify mutation saveDefaultImageForWebsite
* Add defaultImage, images, mediaGroupId, media, and mediaGroupId to ItemMetadata
* Added "root" element to ItemMetadata that points to the root of the collection. Null if the item is the root of the collection.

multimedia-assets was modified to reference marbleContentBucketName, transitioning from rBSCS3ImageBucketName. Note: this will use existing buckets testlibnd-smb-test and libnd-smb-marble. Because of using buckets not created in the CDK stack, CDK is not able to modify existing bucket definitions. Because of this, the bucket policy and CORS will need to be manually maintained, along with the storage gateway, which is manually maintained.

We will use this new marbleContentBucketName bucket to store all marble content, including images and multimedia. The CloudFront Web Distribution will only expose content stored in the folder named "/public-access".