Update service level objectives
We have a new set of agreed upon objectives as of 7/8/21. This updates
all SLOs to match those values. This required bumping the dependencies
to pull in the newer AppSync SLOs from ndlib-cdk and exposing the apiId
from the maintain-metadata-stack.

Additionally, now that we have new agreed upon objectives that should
be achievable, this will also re-enable all alarms so that we get
notifications when things are out of band.