Change static host to support additional aliases
This is primarily to change the static host stack to support provisioning
additional aliases on the cloudfront so that both `marble.nd.edu` and
`marble.library.nd.edu` work.

To get this to work, I needed the ability to pass context overrides with
a list of values (csv), ex: `-c "additionalAliases=abc.nd.edu, xyz.nd.edu"`.
I've added a new mapContextToProps to better handle the deserialization
of prop types from context. I could not figure out a great way to
determine the type of the prop that the value is being applied to (since
these are interfaces), so the caller will just have to specify this with
type hints on each call.

As an example, this call:
`mapContextToProps<IStaticHostStackProps>('website', defaults, { additionalAliases: 'csv' })`
will:
1. Initialize a IStaticHostStackProps object with the passed defaults
2. Glob all context keys with a `website:` prefix
3. Override the given defaults with the globbed values, parsing any
additionalAliases from csv to an array.

Type hints can also be used to convert boolean strings from context to
booleans.

This change also improves the type checking on any values passed as
defaults before context overrides are applied. Given that context is
something still read in at runtime, I could not improve the type checking
on it, but this at least improves the type checking on any values passed
as defaults.

With that added, the rest of the changes are related to getting the website
pipeline to add a `-c "additionalAliases=marble.library.nd.edu"` specifically
for the prod stack only.