Add detailed dashboards
This adds detailed dashboards for all serverless APIs and CloudFront
based CDNs. This additionally changes the manifest lambda apis to enable
detailed metrics. Without this, we won't get metrics per method in the
dashboards.

This does not yet add dashboards for the ElasticSearch and Graphql APIs,
since those do not exist in ndlib-cdk just yet. I've created placeholder
code for now to give an example of what that might look like once it's
added.