Downgrade Marble ElasticSearch back to original settings
Made the following changes so the ElasticSearch stack will again deploy:
* Actually downgraded back to ElasticSearch 7.7
* Changed back from 3 nodes to 2 nodes (in production)
* Changed back from 't2.medium.elasticsearch' to 't2.small.elasticsearch'

Ultimately, to change these settings, we need to let CloudFormation define the name of the ElasticSearch domain, but we can't because of synthetics.

Instead, our solution will be to create a separate OpenSearch stack, and deploy that in parallel to ElasticSearch.  Once we update the static host sites to use OpenSearch, we can decommission the ElasticSearch stack.